Instance Upgrade
Upgrade a trove instance to a new image (presumably with
updated os/guestagent/db) using the Nova rebuild command.

Implements "trove upgrade" CLI command to upgrade an
instance to a new datastore version.

Piggybacked on the instance.edit api, calls upgrade
in the taskmanager, which in turn calls pre_upgrade and
post_upgrade in the guestagent.

Change-Id: I76de869cd3850e297109f09b3260ea9d82460e1f
Implements: blueprint image-upgrade