Add command to delete BUILD instances and clusters
Sometimes an instance/cluster can be stuck in BUILD state forever.
Attempting to delete the instance in this state is currently not
allowed. Add force-delete and reset-status command. Reset-status
will reset the status to ERROR, letting the instance/cluster be
deleted. Force-delete will combine functionality of reset-status
and delete.

Partial-Bug: #1579141

Conflicts:
	troveclient/tests/test_v1_shell.py
	troveclient/v1/shell.py

Change-Id: Ie3b0c1c82e3058630f0ad7dd4bf7db7505af95e9