Add --roles to user-create
A new field for users is being added: roles. This will be a list of
dictionaries sent to the api server. Each role item is expected to be a
dictionary with a mandatory 'name' and optional 'database' keys. The CLI
output will have a new column Roles that will display the list of
roles formatted similarly to the Databases column. If the role has a
'databases' key then the value will be displayed as database:role.

Change-Id: I229c4df3517c78831123a3540f3ac3ccc19a1003
Partial-Bug: dbaas-2160