[CLI] Add Couchbase-specific user properties
Added Couchbase specific properties on user commands.
The user_list code has been updated to handle different
fields returned by different datastores.
The only 'fixed' field is the 'name' which is always
assumed to exist. The remaining properties are filed
dynamically from the received view object.

A side effect of this change is that an empty list
now only displays the 'name' column.

Change-Id: Ib0cecb4dfc58d043198cf0e84782e1ffd23c2bdc
Closes-Bug: DBAAS-2156