Restrict reset-status command to admin
The reset status command should only be used by an Admin as
resetting the state of an instance/cluster can be dangerous.
Any user should still be able to force delete an instance/cluster.

Since, force delete calls the reset status command and then the
delete command it is necessary to pass a flag indicating that
the reset status is being performed for a force delete and the
policy should allow it. If reset status command is being called
for a force delete the body will be appended to indicate it.

Change-Id: I236f6a97fe606c4f03f7ae3b6110d39498b76962
Partial-Bug: DBAAS-2345